fixed up admin UI and made it work when not logged in

+++ src/routes/[lang]/sentences/[id]/delete/+server.ts
@@ -1,11 +1,15 @@
-import { ServerLoad, redirect } from '@sveltejs/kit';
+import { ServerLoad, error, redirect } from '@sveltejs/kit';
 import { deleteSentence } from '../../../../../db/sentences';
 
-export const GET: ServerLoad = async ({ params, locals: { language } }) => {
+export const GET: ServerLoad = async ({ params, locals: { language, isAdmin } }) => {
 	const sentenceId = parseInt(params.id || '');
 
 	if (isNaN(sentenceId)) {
 		throw new Error('Invalid sentence ID');
+	}
+
+	if (!isAdmin) {
+		return error(403, 'Forbidden');
 	}
 
 	await deleteSentence(sentenceId, language);
