+++ src/routes/[lang]/words/[id]/delete/lemma/[word]/+server.ts
@@ -1,11 +1,15 @@
-import { ServerLoad, redirect } from '@sveltejs/kit';
+import { ServerLoad, error, redirect } from '@sveltejs/kit';
 import { deleteWordToLemma } from '../../../../../../../db/lemmas';
 
-export const GET: ServerLoad = async ({ params, locals: { language } }) => {
+export const GET: ServerLoad = async ({ params, locals: { language, isAdmin } }) => {
 	const wordId = parseInt(params.id || '');
 
 	if (isNaN(wordId)) {
 		throw new Error('Invalid word ID');
+	}
+
+	if (!isAdmin) {
+		return error(403, 'Forbidden');
 	}
 
 	await deleteWordToLemma(wordId, params.word!, language);
